[Test Suite] Refactor arithmetic
