+++ test-suite/src/arithmetic/mod.rs
@@ -1,0 +1,3 @@
+pub mod blend;
+pub mod error;
+pub mod on_where;
